Improve Keycloak user validation to restrict access to Che (#173)
Improve Keycloak user validation to restrict access to Che

Update Keycloak integration to check if a provided token matches the
token assigned to the current namespace owner. Updates the
KeycloakAuthenticationFilter, and adds a new tomcat valve
(UserAuthValve) to perform this work.

Adds a new endpoint, /api/token/user, where get requests can be made
to check if the above condition is true.

As of this commit, wsmaster and wsagent have this authentication
enabled. The IDE and dashboard, however, do not.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>