Massive simplification of the Keycloak integration...
...consisting of the following main changes:

- The management of the redirect to the login page + retrieval of the
Keycloak token *only on the client-side* in Javascript (in both the
Dashboard and the GWT applications)
- For the the GWT application, Keycloak initialization + token injection
are now managed inside GWT through promise chaining in the
`KeycloakAsyncRequestFactory` and `KeycloakAsyncRequest` objects.
- In both the Dashboard and the GWT client web applications, the
hard-wired `keycloak.js` files added inside the web application is
replaced by pulling it from the Keycloak server during Keycloak
initialization. This ensures that the Javascript Keycloak adapter code
is always in-sync with the version of the Keycloak server being used.
- All the server-side Java Keycloak adapters (both the tomcat valve
*and* the Keycloak standrad servlet filter) were removed ...
- ... and replaced by our own servlet filter that simply decodes the
received token as a JWT token and checks its signature with the public
key of the Keycloak realm (directly pulled from the Keycloak server)

Signed-off-by: David Festal <dfestal@redhat.com>